+++ src/Drivers/MarkdownDriver.php
@@ -47,7 +47,7 @@ final readonly class MarkdownDriver implements DriverContract
         $content = isset($data['content']) && is_string($data['content']) ? $data['content'] : '';
         unset($data['content'], $data['slug']);
 
-        $yaml = Yaml::dump($data);
+        $yaml = Yaml::dump($data, PHP_INT_MAX);
 
         return "---\n$yaml---\n\n$content\n";
     }